feat: modularize git and GPT operations
### Added

- **`src/git_ops.rs`**: Implemented multiple utility functions for Git operations, including functions to ensure the environment is a Git repository, handle detached HEAD states, change to the repository root, diff management, branch fetching, committing, and creating pull requests via the GitHub CLI.

- **`src/gpt_ops.rs`**: Asynchronous operations for OpenAI APIs to generate branch names and commit descriptions based on a diff context are implemented.

- **`src/tui.rs`**: Defined a text-based UI module using Ratatui to display logs, progress, and application status, with interfaces to render messages and manage user navigation.

### Modified

- **`src/main.rs`**: Refactored the main logic to utilize the newly created modules for Git and GPT operations, as well as the TUI. This includes initializing and updating application logs, managing branch logic, fetching and displaying diffs, and interfacing with GPT for automated descriptions.
